VP1L Unity Assets Added for Import
In this .unitypackage file is all our unity assets with the collision maps I put with them. This is the first rendition, so the file ends with a 1. Most likely this will need revision, perhaps with the sensor I decided to use complex collision mapping on (sorry future me or someone else).