Added base scenes for Lab 4, 5, & 9.
Here are some basic labs I created to build off of. It's three of the six labs we are doing, and two of the labs not completed will be the exact same environment. For those (the pendulum and the other two), the rope asset is required. This is just a start, and once we get this working (such as the projectile gun) I can start making them look better presentation wise.